+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+  * Lazy creation of AWS client.
+  * Add AWS client options to the main configuration for those who want them.
+
 ## 0.1.1
 
   * Update package versions, particularly ESLint due to a rule syntax change.